made refactoring code
