+++ bin/server.js
@@ -4,11 +4,15 @@ const { connectMongo } = require("../model/contacts");
 const PORT = process.env.PORT || 3001;
 
 const start = async () => {
-    await connectMongo();
+    try {
+        await connectMongo();
 
-    app.listen(PORT, () => {
-        console.log(`Server running. Use our API on port: ${PORT}`);
-    });
+        app.listen(PORT, () => {
+            console.log(`Server running. Use our API on port: ${PORT}`);
+        });
+    } catch (err) {
+        console.error(`Failed to launch application with error ${err.message}`);
+    }
 };
 
 start();
