+++ controllers/contacts/getContactControler.js
@@ -1,8 +1,8 @@
-const contactsOperation = require("../../model/contacts");
+const contactsOperation = require("../../services/contacts");
 
 const getContactControler = async (req, res) => {
     const contacts = await contactsOperation.listContacts();
-    return res.json(200, { status: "success", message: contacts });
+    return res.json({ status: "success", message: contacts });
 };
 
 module.exports = getContactControler;
