+++ controllers/contacts/patchContactByIdControler.js
@@ -1,13 +1,17 @@
-const contactsOperation = require("../../model/contacts");
+const contactsOperation = require("../../services/contacts");
 const helpersError = require("../../helpers/responseData");
 
 const patchContactByIdControler = async (req, res) => {
     const contactId = req.params.contactId;
 
     if (!contactId) {
-        return res.sattus9400.json(
-            helpersError.badRequestError("contactId is a required parameter")
-        );
+        return res
+            .sattus(400)
+            .json(
+                helpersError.badRequestError(
+                    "contactId is a required parameter"
+                )
+            );
     }
 
     const updateContact = await contactsOperation.patchContact(
